Se ajustan clases a uso de matriz y operaciones de tipo long

+++ Fuentes/Back-End/BR/BrInterfaces/IOperacion.cs
@@ -4,8 +4,8 @@ namespace Challenge.CubeSummationNS.BR
 {
     public interface IOperacion
     {
-        RespuestaGeneral ProcesarOperacionUpdate(OperacionCubo operacion, int[,,] matriz, ref RespuestaGeneral respuestaGeneral);
+        RespuestaGeneral ProcesarOperacionUpdate(OperacionCubo operacion, long[,,] matriz, ref RespuestaGeneral respuestaGeneral);
 
-        RespuestaGeneral ProcesarOperacionQuery(OperacionCubo operacion, int[,,] matriz, ref RespuestaGeneral respuestaGeneral);
+        RespuestaGeneral ProcesarOperacionQuery(OperacionCubo operacion, long[,,] matriz, ref RespuestaGeneral respuestaGeneral);
     }
 }
